requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=hd77b12b_0
altair=5.0.1=py311haa95532_0
arrow-cpp=11.0.0=ha81ea56_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py311haa95532_0
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
blinker=1.4=py311haa95532_0
bokeh=2.4.3=pypi_0
boost-cpp=1.73.0=h2bbff1b_12
bottleneck=1.3.5=py311h5bb9823_0
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
c-ares=1.19.0=h2bbff1b_0
ca-certificates=2023.7.22=h56e8100_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311haa95532_0
colorama=0.4.6=py311haa95532_0
comm=0.1.2=py311haa95532_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py311h21b164f_2
cycler=0.11.0=pypi_0
debugpy=1.5.1=py311hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
docopt=0.6.2=py_1
executing=0.8.3=pyhd3eb1b0_0
faker=19.2.0=pypi_0
favicon=0.7.0=pypi_0
fonttools=4.41.1=pypi_0
freetype=2.12.1=ha860e81_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py311haa95532_0
glog=0.5.0=hd77b12b_0
grpc-cpp=1.48.2=hfe90ff0_1
htbuilder=0.6.1=pypi_0
idna=3.4=py311haa95532_0
importlib-metadata=6.0.0=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.19.2=py311h86cfffd_0
ipython=8.12.0=py311haa95532_0
ipywidgets=8.0.4=py311haa95532_0
jedi=0.18.1=py311haa95532_1
jinja2=3.1.2=py311haa95532_0
joblib=1.3.1=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py311haa95532_0
jupyter_client=8.1.0=py311haa95532_0
jupyter_core=5.3.0=py311haa95532_0
jupyterlab_widgets=3.0.5=py311haa95532_0
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_12
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcurl=8.1.1=h86230a5_0
libdeflate=1.17=h2bbff1b_0
libevent=2.1.12=h56d1f94_1
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.0=h6c2663c_2
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lxml=4.9.3=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=py311haa95532_1
markdownlit=0.0.7=pypi_0
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=py311haa95532_0
mdurl=0.1.0=py311haa95532_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
more-itertools=10.0.0=pypi_0
nest-asyncio=1.5.6=py311haa95532_0
numexpr=2.8.4=py311h1fcbade_1
numpy=1.25.0=py311hdab7c0b_0
numpy-base=1.25.0=py311hd01c5d8_0
openssl=3.0.9=h2bbff1b_0
orc=1.7.4=h623e30f_1
packaging=23.0=py311haa95532_0
pandas=1.5.3=py311heda8569_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py311hd77b12b_0
pip=23.1.2=py311haa95532_0
pipreqs=0.4.13=pyhd8ed1ab_0
platformdirs=2.5.2=py311haa95532_0
prompt-toolkit=3.0.36=py311haa95532_0
protobuf=3.20.3=py311hd77b12b_0
psutil=5.9.0=py311h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=py311h8a3a540_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=pypi_0
pygments=2.15.1=py311haa95532_1
pymdown-extensions=10.1=pypi_0
pympler=0.9=py_0
pyopenssl=23.0.0=py311haa95532_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py311haa95532_0
pywin32=305=py311h2bbff1b_0
pyyaml=6.0=py311h2bbff1b_1
pyzmq=25.1.0=py311hd77b12b_0
re2=2022.04.01=hd77b12b_0
requests=2.29.0=py311haa95532_0
rich=13.3.5=py311haa95532_0
scikit-learn=1.2.2=pypi_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h2bbff1b_0
st-annotated-text=4.0.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.25.0=pypi_0
streamlit-camera-input-live=0.2.0=pypi_0
streamlit-card=0.0.61=pypi_0
streamlit-embedcode=0.1.2=pypi_0
streamlit-extras=0.2.7=pypi_0
streamlit-faker=0.0.2=pypi_0
streamlit-image-coordinates=0.1.6=pypi_0
streamlit-keyup=0.2.0=pypi_0
streamlit-toggle-switch=1.0.2=pypi_0
streamlit-vertical-slider=1.0.2=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=py311haa95532_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py311haa95532_0
tornado=6.2=py311h2bbff1b_0
traitlets=5.7.1=py311haa95532_0
typing_extensions=4.6.3=py311haa95532_0
tzdata=2023c=h04d1e81_0
tzlocal=2.1=py311haa95532_1
urllib3=1.26.16=py311haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py311haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py311haa95532_0
widgetsnbextension=4.0.5=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xyzservices=2023.7.0=pypi_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
yarg=0.1.9=py_1
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py311haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
